Add Dentist class and AddDentistCommand (Part 4)
Let's Implement the Logic component to handle AddDentistCommand
Changes made:
* Logic and Logic Manager files include methods for dentist objects
* Bug fix for AddressBookParser as it was not parsing AddDentistCommand

Note: This commit produced a working version of the app. There is a message indicating successful
execution of add-dentist command. However, the dentist added is still not stored
in the database because Storage component has yet to be modified.

The next step would be to modify Storage component to store dentist datas.